+++ src/stream/mod.rs
@@ -270,19 +270,36 @@ pub trait Range {
 }
 
 impl<'a> RangeStreamOnce for &'a str {
-    #[inline]
     fn uncons_while<F>(&mut self, mut f: F) -> Result<&'a str, StreamErrorFor<Self>>
     where
         F: FnMut(Self::Item) -> bool,
     {
         let mut chars = self.chars();
-        while let Some(c) = chars.next() {
-            if !f(c) {
-                let len = self.len() - chars.as_str().len() - c.len_utf8();
-                let (result, rest) = self.split_at(len);
-                *self = rest;
-                return Ok(result);
+
+        macro_rules! test_next {
+            () => {
+                match chars.next() {
+                    Some(c) => {
+                        if !f(c) {
+                            let len = self.len() - chars.as_str().len() - c.len_utf8();
+                            let (result, rest) = self.split_at(len);
+                            *self = rest;
+                            return Ok(result);
+                        }
+                    }
+                    None => break,
+                }
             }
+        }
+        loop {
+            test_next!();
+            test_next!();
+            test_next!();
+            test_next!();
+            test_next!();
+            test_next!();
+            test_next!();
+            test_next!();
         }
         let result = *self;
         *self = &self[self.len()..];
@@ -617,8 +634,11 @@ where
     ///
     /// NOTE: This type do not implement `Positioned` and `Clone` and must be wrapped with types
     ///     such as `BufferedStreamRef` and `State` to become a `Stream` which can be parsed
-    pub fn new(iter: I) -> IteratorStream<I> {
-        IteratorStream(iter)
+    pub fn new<T>(iter: T) -> IteratorStream<I>
+    where
+        T: IntoIterator<IntoIter = I, Item = I::Item>,
+    {
+        IteratorStream(iter.into_iter())
     }
 }